feat: ask user for confirmation when purging a rock tree

+++ rocks-bin/src/purge.rs
@@ -1,4 +1,5 @@
 use eyre::Result;
+use inquire::Confirm;
 use rocks_lib::{config::Config, tree::Tree};
 
 pub fn purge(config: Config) -> Result<()> {
@@ -7,7 +8,14 @@ pub fn purge(config: Config) -> Result<()> {
         config.lua_version().cloned().unwrap(),
     )?;
 
-    std::fs::remove_dir_all(tree.root())?;
+    let len = tree.list().len();
+
+    if Confirm::new(&format!("Are you sure you want to purge all {len} rocks?"))
+        .with_default(false)
+        .prompt()?
+    {
+        std::fs::remove_dir_all(tree.root())?;
+    }
 
     Ok(())
 }
